requirements.txt
aniso8601==9.0.1
apispec==5.1.1
azure-core==1.19.0
azure-storage-blob==12.9.0
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.6
click==8.0.1
colorama==0.4.4
cryptography==35.0.0
cycler==0.10.0
Flask==2.0.2
flask-apispec==0.11.0
Flask-Cors==3.0.10
Flask-PyMongo==2.3.0
Flask-RESTful==0.3.9
idna==3.2
imageio==2.9.0
isodate==0.6.0
itsdangerous==2.0.1
Jinja2==3.0.1
joblib==1.0.1
kiwisolver==1.3.2
MarkupSafe==2.0.1
marshmallow==3.13.0
matplotlib==3.4.3
msrest==0.6.21
networkx==2.6.3
numpy==1.21.2
oauthlib==3.1.1
opencv-python-headless==4.5.3.56
Pillow==8.3.2
pycparser==2.20
PyJWT==2.1.0
pymongo==3.12.0
pyparsing==2.4.7
python-dateutil==2.8.2
pytz==2021.1
PyWavelets==1.1.1
requests==2.26.0
requests-oauthlib==1.3.0
scikit-image==0.18.3
scikit-learn==1.0
scipy==1.7.1
six==1.16.0
sklearn==0.0
threadpoolctl==2.2.0
tifffile==2021.8.30
urllib3==1.26.7
webargs==8.0.1
Werkzeug==2.0.1
